% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print function}
\usage{
hello(message = "Hello World!", print.message = TRUE)
}
\arguments{
\item{message}{a character string (default 'Hello World!')}

\item{print.message}{the message is printed if TRUE (default TRUE)}
}
\value{
number of characters in the message.
}
\description{
This function prints out a message and returns the number of characters in the message.
}
\examples{
hello()

hello('hahaha')

hello('lalala', FALSE)

}
\keyword{greeting}
