% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_build_index.R
\name{px_build_index}
\alias{px_build_index}
\title{Check function on pairix-indexed pairs file.}
\usage{
px_build_index(filename, preset = "pairs", force = FALSE)
}
\arguments{
\item{filename}{a pairs file, or a bgzipped text file (sometextfile.gz) with an index file sometextfile.gz.px2 in the same folder.}

\item{preset}{one of the following strings: 'gff', 'bed', 'sam', 'vcf', 'psltbl' (1D-indexing) or 'pairs', 'merged_nodups', 'old_merged_nodups' (2D-indexing). (default 'pairs')}

\item{force}{If TRUE, overwrite existing index file. If FALSE, do not overwrite unless the index file is older than the bgzipped file. (default FALSE)}
}
\description{
This function allows you to check if a key (chr for 1D, chr pair for 2D) exists in a pairs file.
}
\examples{

filename = system.file(".","test_4dn.pairs.gz", package="Rpairix")
px_build_index(filename)

}
\keyword{index}
\keyword{pairix}

