% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_colnames.R
\name{px_colnames}
\alias{px_colnames}
\title{Column name retrival function on pairix-indexed pairs file.}
\usage{
px_colnames(filename)
}
\arguments{
\item{filename}{a pairs file, or a bgzipped text file (sometextfile.gz) with an index file sometextfile.gz.px2 in the same folder.}
}
\description{
This function returns a vector of column names for a pairs format.
}
\examples{
filename = system.file(".","test_4dn.pairs.gz", package="Rpairix")
cols = px_colnames(filename)
print(cols)
}
\keyword{names}
\keyword{pairix}

