% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_exists.R
\name{px_exists}
\alias{px_exists}
\title{Check function on pairix-indexed pairs file.}
\usage{
px_exists(filename, key)
}
\arguments{
\item{filename}{a pairs file, or a bgzipped text file (sometextfile.gz) with an index file sometextfile.gz.px2 in the same folder.}

\item{key}{a pair of chromosomes in the query string format (e.g. "chr1|chr2"), or a chromosome for a 1D-indexed pairs file (e.g. "chr1").}
}
\value{
TRUE if the key exists or FALSE if not. If index loading fails, NULL is returned.
}
\description{
This function allows you to check if a key (chr for 1D, chr pair for 2D) exists in a pairs file.
}
\examples{
filename = system.file(".","test_4dn.pairs.gz", package="Rpairix")
key = "chrX|chrX"
res = px_exists(filename, key)
print(res)

filename = system.file(".","merged_nodups.space.chrblock_sorted.subsample1.txt.gz",package="Rpairix")
key = "10|20"
res = px_exists(filename, key)
print(res)

filename = system.file(".","merged_nodups.space.chrblock_sorted.subsample1.txt.gz",
package="Rpairix")
key = "10|20"
res = px_exists(filename, key)
print(res)
}
\keyword{check}
\keyword{pairix}

