% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_exists2.R
\name{px_exists2}
\alias{px_exists2}
\title{Check function on pairix-indexed 2D pairs file.}
\usage{
px_exists2(filename, chr1, chr2)
}
\arguments{
\item{filename}{a pairs file, or a bgzipped text file (sometextfile.gz) with an index file sometextfile.gz.px2 in the same folder.}

\item{chr1}{first chromosome}

\item{chr2}{second chromosome}
}
\description{
This function allows you to check if a pair of chromosomes exists in a 2D-indexed pairs file.
}
\examples{
filename = system.file(".","test_4dn.pairs.gz", package="Rpairix")
chr1 = "chr22"
chr2 = "chrX"
res = px_exists2(filename, chr1, chr2)
print(res)

filename = system.file(".","merged_nodups.space.chrblock_sorted.subsample1.txt.gz",package="Rpairix")
chr1 = "10"
chr2 = "20"
res = px_exists2(filename, chr1, chr2)
print(res)

filename = system.file(".","merged_nodups.space.chrblock_sorted.subsample1.txt.gz",
package="Rpairix")
chr1 = "10"
chr2 = "20"
res = px_exists2(filename, chr1, chr2)
print(res)
}
\keyword{check}
\keyword{pairix}

