% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_get_linecount.R
\name{px_get_linecount}
\alias{px_get_linecount}
\title{Function on pairix-indexed pairs file.}
\usage{
px_get_linecount(filename)
}
\arguments{
\item{filename}{a pairs file, or a bgzipped text file (sometextfile.gz) with an index file sometextfile.gz.px2 in the same folder.}
}
\value{
An integer corresponding to total line count, NULL if error.
}
\description{
This function returns the total line count of a pairs file (equivalent to gunzip -c | wc -l but much faster)
}
\examples{
filename = system.file(".","test_4dn.pairs.gz", package="Rpairix")
px_get_linecount(filename)

filename = system.file(".","merged_nodups.space.chrblock_sorted.subsample1.txt.gz",package="Rpairix")
px_get_linecount(filename)

filename = system.file(".","merged_nodups.space.chrblock_sorted.subsample1.txt.gz",
package="Rpairix")
px_get_linecount(filename)

}
\keyword{count}
\keyword{line}
\keyword{linecount}
\keyword{pairix}
