% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_getChar.R
\name{px_getChar}
\alias{px_getChar}
\title{Test function}
\usage{
px_getChar(char_vector)
}
\arguments{
\item{char_vector}{a character vector}
}
\description{
This function is a test function.
}
\keyword{pairix}
\keyword{test}

