% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_query_gr.R
\name{px_query_gr}
\alias{px_query_gr}
\title{Query pairix-indexed pairs file using GenomicRanges-related objects.}
\usage{
px_query_gr(filename, queryobj, ...)
}
\arguments{
\item{filename}{a pairs file, or a bgzipped text file (sometextfile.gz) with an index file sometextfile.gz.px2 in the same folder.}

\item{queryobj}{One of three types: (1) a character vector containing a set of pairs of genomic coordinates in 1-based "chr1:start1-end1|chr2:start2-end2" format. start-end can be omitted (e.g. "chr1:start1-end1|chr2" or "chr1|chr2"); (2) A GInteractions object from the package "InteractionSet"; (3) A GRangesList composed of two GRanges objects of identical length (first pairs, second pairs).}

\item{...}{Any of the other parameters from px_query(), such as max_mem, stringsAsFactors, linecount.only, autoflip.}
}
\value{
data frame containing the query result. Column names are added if indexing was done with a pairs preset.
}
\description{
This function allows you to query a 2D range in a pairix-indexed pairs file from several query types.
}
\details{
This function requires Bioconductor packages InteractionSet and GenomicRanges.
}
\examples{

filename = system.file(".","test_4dn.pairs.gz", package="Rpairix")

# -- construct query objects -- #

# GenomicRanges::GRangesList
gr <- GRanges(
  seqnames = Rle(c("chr10", "chr20", "chr21", "chr22"), c(1, 2, 1, 2)),
  ranges = IRanges((0:5*1000000)+1, end = (0:5*1000000)+13000000))
grl <- split(gr, rep(1:2,3))
grl

# InteractionSet::GInteractions
gi <- GInteractions(grl[[1]],grl[[2]])
gi

# -- query -- #

# with GInteractions
res = px_query_gr(filename,queryobj=gi)
print(res)

# with GRangesList
res = px_query_gr(filename,queryobj=grl)
print(res)

}
\keyword{2D}
\keyword{GInteractions}
\keyword{GenomicRanges}
\keyword{pairix}
\keyword{query}
