% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_seq1list.R
\name{px_seq1list}
\alias{px_seq1list}
\title{Function to get the list of first chromosomes from a pairix-indexed pairs file.}
\usage{
px_seq1list(filename)
}
\arguments{
\item{filename}{a pairs file, or a bgzipped text file (sometextfile.gz) with an index file sometextfile.gz.px2 in the same folder.}
}
\description{
This function allows you to get the list of first chromosomes in a pairix-indexed pairs file.
}
\examples{
filename = system.file(".","test_4dn.pairs.gz", package="Rpairix")
res = px_seq1list(filename)
print(res)

filename = system.file(".","merged_nodups.space.chrblock_sorted.subsample1.txt.gz",package="Rpairix")
res = px_seq1list(filename)
print(res)
}
\keyword{2D}
\keyword{pairix}
\keyword{query}

