% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_seq2list.R
\name{px_seq2list}
\alias{px_seq2list}
\title{Function to get the list of second chromosomes from a pairix-indexed pairs file.}
\usage{
px_seq2list(filename)
}
\arguments{
\item{filename}{a pairs file, or a bgzipped text file (sometextfile.gz) with an index file sometextfile.gz.px2 in the same folder.}
}
\description{
This function allows you to get the list of second chromosomes in a pairix-indexed pairs file.
}
\examples{
filename = "merged_nodup.tab.chrblock_sorted.txt.gz"
res = px_seq2list(filename)
}
\keyword{2D}
\keyword{pairix}
\keyword{query}

