% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_build_index.R
\name{px_build_index}
\alias{px_build_index}
\title{Indexing function on bgzipped pairs file.}
\usage{
px_build_index(filename, preset = "", sc = 0, bc = 0, ec = 0, sc2 = 0,
  bc2 = 0, ec2 = 0, delimiter = "\\t", comment_char = "#",
  region_split_character = "|", line_skip = 0, force = FALSE)
}
\arguments{
\item{filename}{a pairs file, or a bgzipped text file (sometextfile.gz) with an index file sometextfile.gz.px2 in the same folder.}

\item{preset}{one of the following strings: 'gff', 'bed', 'sam', 'vcf', 'psltbl' (1D-indexing) or 'pairs', 'merged_nodups', 'old_merged_nodups' (2D-indexing). If preset is '', at least some of the custom parameters must be given instead (sc, bc, ec, sc2, bc2, ec2, delimiter, comment_char, line_skip). (default '')}

\item{sc}{first sequence (chromosome) column index (1-based). Zero (0) means not specified. If preset is given, preset overrides sc. If preset is not given, this one is required. (default 0)}

\item{bc}{first start position column index (1-based). Zero (0) means not specified. If preset is given, preset overrides bc. If preset is not given, this one is required. (default 0)}

\item{ec}{first end position column index (1-based). Zero (0) means not specified. If preset is given, preset overrides ec. (default 0)}

\item{sc2}{second sequence (chromosome) column index (1-based). Zero (0) means not specified. If preset is given, preset overrides sc2. If sc, bc are specified but not sc2 and bc2, it is 1D-indexed. (default 0)}

\item{bc2}{second start position column index (1-based). Zero (0) means not specified. If preset is given, preset overrides bc2. (default 0)}

\item{ec2}{second end position column index (1-based). Zero (0) means not specified. If preset is given, preset overrides ec2. (default 0)}

\item{delimiter}{delimiter (e.g. '\\t' or ' ') (default '\\t'). If preset is given, preset overrides delimiter.}

\item{comment_char}{comment character. Lines beginning with this character are skipped when creating an index. If preset is given, preset overrides comment_char (default '#')}

\item{region_split_character}{region_split_character (default '|'). This option overrides preset. (All presets have default region_split_character ('|')). This parameter can be useful when one's chromosome names contain character '|'.}

\item{line_skip}{number of lines to skip in the beginning. (default 0)}

\item{force}{If TRUE, overwrite existing index file. If FALSE, do not overwrite unless the index file is older than the bgzipped file. (default FALSE)}
}
\description{
This function creates a pairix (px2) index a bgzipped text file. Either a preset or a set of custom parameters (column indices, comment_char, line_skip) must be specified.
}
\examples{

filename = system.file(".","test_4dn.pairs.gz", package="Rpairix")
px_build_index(filename, force=TRUE)
px_query(filename, 'chr22|chr22')
px_build_index(filename, 'pairs', force=TRUE)
px_build_index(filename, sc=2, bc=3, ec=3, sc2=4, bc2=5, ec2=5, force=TRUE)
px_build_index(filename, region_split_character='^', force=TRUE)
px_query(filename, 'chr22^chr22')

}
\keyword{index}
\keyword{pairix}
